Revert "refactor: fluent builder pattern for generated classes"
This reverts commit 88cd8ca94f47cde8bafe3fd31c6b910726b770fb.

+++ src/main/java/com/sipgate/li/lib/x1/protocol/error/GenericBaseException.java
@@ -1,5 +1,6 @@
 package com.sipgate.li.lib.x1.protocol.error;
 
+import java.math.BigInteger;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -7,7 +8,7 @@ abstract class GenericBaseException extends ErrorResponseException {
 
   private static final Logger LOGGER = LoggerFactory.getLogger(GenericBaseException.class);
 
-  protected GenericBaseException(final long errorCode, final Exception e) {
+  protected GenericBaseException(final BigInteger errorCode, final Exception e) {
     // don't expose possibly confidential information from generic exception
     super(errorCode);
     // but write everything to the log so developers can debug
