+++ src/test/java/com/sipgate/li/lib/x1/protocol/error/ErrorResponseExceptionTest.java
@@ -3,6 +3,7 @@ package com.sipgate.li.lib.x1.protocol.error;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.lang.reflect.Modifier;
+import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.HashSet;
 import org.junit.jupiter.api.Test;
@@ -14,14 +15,14 @@ class ErrorResponseExceptionTest {
   void itHasUniqueErrorCodes() throws IllegalAccessException {
     final var fields = ReflectionUtils.getFields(
       ErrorResponseException.class,
-      field -> field.getType() == Integer.class,
+      field -> field.getType() == BigInteger.class,
       field -> Modifier.isStatic(field.getModifiers()),
       field -> Modifier.isFinal(field.getModifiers())
     );
 
-    final var allErrorCodes = new ArrayList<Integer>();
+    final var allErrorCodes = new ArrayList<BigInteger>();
     for (final var field : fields) {
-      allErrorCodes.add(((Integer) field.get(null)));
+      allErrorCodes.add(((BigInteger) field.get(null)));
     }
 
     final var uniqueErrorCodes = new HashSet<>(allErrorCodes);
